requirements.txt
requests==2.31.0
python-dotenv==1.0.1
openai==1.14.1
google-cloud-texttospeech==2.14.1
google-cloud-speech
pyTelegramBotAPI==4.14.1
flask
